requirements.txt
streamlit
phidata
yfinance
duckduckgo-search